Revert "Change get_message_ID to return a uint16_t"
This reverts commit d34278b8857a0974815d0d97e206e17690df9f13. It would
not have worked with the MotorControllerFrameRequest struct.

+++ Common/lib/CANStructs/CANStruct.h
@@ -10,7 +10,8 @@
 struct CANStruct {
     virtual void serialize(CANMessage *message);
     virtual void deserialize(CANMessage *message);
-    virtual uint16_t get_message_ID();
+    // TODO: Change to uint16_t
+    virtual uint32_t get_message_ID();
     virtual void print();
 };
